moved actionTypes to constants folder and set getEventReducer to return rootReducer
