+++ src/shared/constants/actionTypes.js
