+++ src/shared/store/actions/index.js
@@ -1,2 +1,1 @@
 export * from "./bookmarkNodesActions";
-export * from "./actionTypes";
